+++ tests/integration/x/vm/test_state_transition.go
@@ -7,8 +7,8 @@ import (
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core"
-	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	gethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/holiman/uint256"
@@ -16,6 +16,7 @@ import (
 	"github.com/cometbft/cometbft/crypto/tmhash"
 	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 
+	rpctypes "github.com/cosmos/evm/rpc/types"
 	"github.com/cosmos/evm/testutil/config"
 	"github.com/cosmos/evm/testutil/integration/evm/factory"
 	"github.com/cosmos/evm/testutil/integration/evm/grpc"
@@ -628,10 +629,10 @@ func (s *KeeperTestSuite) TestApplyTransaction() {
 }
 
 type testHooks struct {
-	postProcessing func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *ethtypes.Receipt) error
+	postProcessing func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *gethtypes.Receipt) error
 }
 
-func (h *testHooks) PostTxProcessing(ctx sdk.Context, sender common.Address, msg core.Message, receipt *ethtypes.Receipt) error {
+func (h *testHooks) PostTxProcessing(ctx sdk.Context, sender common.Address, msg core.Message, receipt *gethtypes.Receipt) error {
 	return h.postProcessing(ctx, sender, msg, receipt)
 }
 
@@ -651,7 +652,7 @@ func (s *KeeperTestSuite) TestApplyTransactionWithTxPostProcessing() {
 				s.Network.App.GetEVMKeeper().SetHooks(
 					keeper.NewMultiEvmHooks(
 						&testHooks{
-							postProcessing: func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *ethtypes.Receipt) error {
+							postProcessing: func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *gethtypes.Receipt) error {
 								return nil
 							},
 						},
@@ -691,7 +692,7 @@ func (s *KeeperTestSuite) TestApplyTransactionWithTxPostProcessing() {
 				s.Network.App.GetEVMKeeper().SetHooks(
 					keeper.NewMultiEvmHooks(
 						&testHooks{
-							postProcessing: func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *ethtypes.Receipt) error {
+							postProcessing: func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *gethtypes.Receipt) error {
 								return errors.New("post processing failed :(")
 							},
 						},
@@ -731,7 +732,7 @@ func (s *KeeperTestSuite) TestApplyTransactionWithTxPostProcessing() {
 				s.Network.App.GetEVMKeeper().SetHooks(
 					keeper.NewMultiEvmHooks(
 						&testHooks{
-							postProcessing: func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *ethtypes.Receipt) error {
+							postProcessing: func(ctx sdk.Context, sender common.Address, msg core.Message, receipt *gethtypes.Receipt) error {
 								return s.Network.App.GetMintKeeper().MintCoins(
 									ctx, sdk.NewCoins(sdk.NewCoin("arandomcoin", sdkmath.NewInt(100))),
 								)
@@ -830,19 +831,32 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 	s.EnableFeemarket = true
 	defer func() { s.EnableFeemarket = false }()
 	s.SetupTest()
+
+	testAddr := utiltx.GenerateAddress()
+	balance := (*hexutil.Big)(big.NewInt(1000000000000000000))
+	nonce := hexutil.Uint64(0)
+
+	overrides := rpctypes.StateOverride{
+		testAddr: rpctypes.OverrideAccount{
+			Balance: &balance,
+			Nonce:   &nonce,
+		},
+	}
+
 	testCases := []struct {
 		name               string
 		getMessage         func() core.Message
 		getEVMParams       func() types.Params
 		getFeeMarketParams func() feemarkettypes.Params
+		overrides          *rpctypes.StateOverride
 		expErr             bool
 		expVMErr           bool
 		expectedGasUsed    uint64
 		postCheck          func()
 	}{
 		{
-			"success - messsage applied ok with default params",
-			func() core.Message {
+			name: "success - messsage applied ok with default params",
+			getMessage: func() core.Message {
 				sender := s.Keyring.GetKey(0)
 				recipient := s.Keyring.GetAddr(1)
 				msg, err := s.Factory.GenerateGethCoreMsg(sender.Priv, types.EvmTxArgs{
@@ -852,23 +866,24 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 				s.Require().NoError(err)
 				return *msg
 			},
-			types.DefaultParams,
-			feemarkettypes.DefaultParams,
-			false,
-			false,
-			params.TxGas,
-			nil,
+			getEVMParams:       types.DefaultParams,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          nil,
+			expErr:             false,
+			expVMErr:           false,
+			expectedGasUsed:    params.TxGas,
+			postCheck:          nil,
 		},
 		{
-			"success - applies set code authorization (EIP-7702)",
-			func() core.Message {
+			name: "success - applies set code authorization (EIP-7702)",
+			getMessage: func() core.Message {
 				authority := s.Keyring.GetKey(0)
 				target := s.Keyring.GetAddr(1)
 
 				accResp, err := s.Handler.GetEvmAccount(authority.Addr)
 				s.Require().NoError(err)
 
-				auth := ethtypes.SetCodeAuthorization{
+				auth := gethtypes.SetCodeAuthorization{
 					ChainID: *uint256.NewInt(types.GetChainConfig().GetChainId()),
 					Address: target,
 					Nonce:   accResp.GetNonce(),
@@ -878,41 +893,42 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 
 				msg, err := s.Factory.GenerateGethCoreMsg(authority.Priv, types.EvmTxArgs{
 					To:                &common.Address{},
-					AuthorizationList: []ethtypes.SetCodeAuthorization{signedAuth},
+					AuthorizationList: []gethtypes.SetCodeAuthorization{signedAuth},
 				})
 				s.Require().NoError(err)
 				return *msg
 			},
-			types.DefaultParams,
-			feemarkettypes.DefaultParams,
-			false,
-			false,
-			params.TxGas + params.CallNewAccountGas -
+			getEVMParams:       types.DefaultParams,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          nil,
+			expErr:             false,
+			expVMErr:           false,
+			expectedGasUsed: params.TxGas + params.CallNewAccountGas -
 				keeper.GasToRefund(
 					params.CallNewAccountGas-params.TxAuthTupleGas,
 					params.TxGas+params.CallNewAccountGas,
 					params.RefundQuotientEIP3529,
 				),
-			func() {
+			postCheck: func() {
 				authorityAddr := s.Keyring.GetAddr(0)
 				targetAddr := s.Keyring.GetAddr(1)
 				codeHash := s.Network.App.GetEVMKeeper().GetCodeHash(s.Network.GetContext(), authorityAddr)
 				code := s.Network.App.GetEVMKeeper().GetCode(s.Network.GetContext(), codeHash)
-				delegationAddr, ok := ethtypes.ParseDelegation(code)
+				delegationAddr, ok := gethtypes.ParseDelegation(code)
 				s.Require().True(ok)
 				s.Require().Equal(targetAddr, delegationAddr)
 			},
 		},
 		{
-			"fail - unsigned set code authorization is ignored (EIP-7702)",
-			func() core.Message {
+			name: "fail - unsigned set code authorization is ignored (EIP-7702)",
+			getMessage: func() core.Message {
 				authority := s.Keyring.GetKey(0)
 				target := s.Keyring.GetAddr(1)
 
 				accResp, err := s.Handler.GetEvmAccount(authority.Addr)
 				s.Require().NoError(err)
 
-				auth := ethtypes.SetCodeAuthorization{
+				auth := gethtypes.SetCodeAuthorization{
 					ChainID: *uint256.NewInt(types.GetChainConfig().GetChainId()),
 					Address: target,
 					Nonce:   accResp.GetNonce(),
@@ -920,28 +936,87 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 
 				msg, err := s.Factory.GenerateGethCoreMsg(authority.Priv, types.EvmTxArgs{
 					To:                &common.Address{},
-					AuthorizationList: []ethtypes.SetCodeAuthorization{auth},
+					AuthorizationList: []gethtypes.SetCodeAuthorization{auth},
 				})
 				s.Require().NoError(err)
 				return *msg
 			},
-			types.DefaultParams,
-			feemarkettypes.DefaultParams,
-			false,
-			false,
-			params.TxGas + params.CallNewAccountGas,
-			func() {
+			getEVMParams:       types.DefaultParams,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          nil,
+			expErr:             false,
+			expVMErr:           false,
+			expectedGasUsed:    params.TxGas + params.CallNewAccountGas,
+			postCheck: func() {
 				authorityAddr := s.Keyring.GetAddr(0)
 				codeHash := s.Network.App.GetEVMKeeper().GetCodeHash(s.Network.GetContext(), authorityAddr)
 				code := s.Network.App.GetEVMKeeper().GetCode(s.Network.GetContext(), codeHash)
-				_, ok := ethtypes.ParseDelegation(code)
+				_, ok := gethtypes.ParseDelegation(code)
 				s.Require().False(ok)
 				s.Require().Len(code, 0)
 			},
 		},
 		{
-			"call contract tx with config param EnableCall = false",
-			func() core.Message {
+			name: "success - message applied with balance override",
+			getMessage: func() core.Message {
+				sender := s.Keyring.GetKey(0)
+				recipient := s.Keyring.GetAddr(1)
+				msg, err := s.Factory.GenerateGethCoreMsg(sender.Priv, types.EvmTxArgs{
+					To:       &recipient,
+					Amount:   big.NewInt(100),
+					GasLimit: params.TxGas,
+				})
+				s.Require().NoError(err)
+				return *msg
+			},
+			getEVMParams:       types.DefaultParams,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          &overrides,
+			expErr:             false,
+			expVMErr:           false,
+			expectedGasUsed:    params.TxGas,
+		},
+		{
+			name: "success - simple transfer from overridden address",
+			getMessage: func() core.Message {
+				sender := s.Keyring.GetKey(0)
+				recipient := s.Keyring.GetAddr(1)
+				msg, err := s.Factory.GenerateGethCoreMsg(sender.Priv, types.EvmTxArgs{
+					To:     &recipient,
+					Amount: big.NewInt(50),
+				})
+				s.Require().NoError(err)
+				return *msg
+			},
+			getEVMParams:       types.DefaultParams,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          &overrides,
+			expErr:             false,
+			expVMErr:           false,
+			expectedGasUsed:    params.TxGas,
+		},
+		{
+			name: "success - empty state overrides",
+			getMessage: func() core.Message {
+				sender := s.Keyring.GetKey(0)
+				recipient := s.Keyring.GetAddr(1)
+				msg, err := s.Factory.GenerateGethCoreMsg(sender.Priv, types.EvmTxArgs{
+					To:     &recipient,
+					Amount: big.NewInt(100),
+				})
+				s.Require().NoError(err)
+				return *msg
+			},
+			getEVMParams:       types.DefaultParams,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          &rpctypes.StateOverride{},
+			expErr:             false,
+			expVMErr:           false,
+			expectedGasUsed:    params.TxGas,
+		},
+		{
+			name: "call contract tx with config param EnableCall = false",
+			getMessage: func() core.Message {
 				sender := s.Keyring.GetKey(0)
 				recipient := s.Keyring.GetAddr(1)
 				msg, err := s.Factory.GenerateGethCoreMsg(sender.Priv, types.EvmTxArgs{
@@ -952,7 +1027,7 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 				s.Require().NoError(err)
 				return *msg
 			},
-			func() types.Params {
+			getEVMParams: func() types.Params {
 				defaultParams := types.DefaultParams()
 				defaultParams.AccessControl = types.AccessControl{
 					Call: types.AccessControlType{
@@ -961,15 +1036,16 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 				}
 				return defaultParams
 			},
-			feemarkettypes.DefaultParams,
-			false,
-			true,
-			0,
-			nil,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          nil,
+			expErr:             false,
+			expVMErr:           true,
+			expectedGasUsed:    0,
+			postCheck:          nil,
 		},
 		{
-			"create contract tx with config param EnableCreate = false",
-			func() core.Message {
+			name: "create contract tx with config param EnableCreate = false",
+			getMessage: func() core.Message {
 				sender := s.Keyring.GetKey(0)
 				msg, err := s.Factory.GenerateGethCoreMsg(sender.Priv, types.EvmTxArgs{
 					Amount: big.NewInt(100),
@@ -978,7 +1054,7 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 				s.Require().NoError(err)
 				return *msg
 			},
-			func() types.Params {
+			getEVMParams: func() types.Params {
 				defaultParams := types.DefaultParams()
 				defaultParams.AccessControl = types.AccessControl{
 					Create: types.AccessControlType{
@@ -987,15 +1063,16 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 				}
 				return defaultParams
 			},
-			feemarkettypes.DefaultParams,
-			false,
-			true,
-			0,
-			nil,
+			getFeeMarketParams: feemarkettypes.DefaultParams,
+			overrides:          nil,
+			expErr:             false,
+			expVMErr:           true,
+			expectedGasUsed:    0,
+			postCheck:          nil,
 		},
 		{
-			"fail - fix panic when minimumGasUsed is not uint64",
-			func() core.Message {
+			name: "fail - fix panic when minimumGasUsed is not uint64",
+			getMessage: func() core.Message {
 				sender := s.Keyring.GetKey(0)
 				recipient := s.Keyring.GetAddr(1)
 				msg, err := s.Factory.GenerateGethCoreMsg(sender.Priv, types.EvmTxArgs{
@@ -1005,18 +1082,19 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 				s.Require().NoError(err)
 				return *msg
 			},
-			types.DefaultParams,
-			func() feemarkettypes.Params {
+			getEVMParams: types.DefaultParams,
+			getFeeMarketParams: func() feemarkettypes.Params {
 				paramsRes, err := s.Handler.GetFeeMarketParams()
 				s.Require().NoError(err)
 				params := paramsRes.GetParams()
 				params.MinGasMultiplier = sdkmath.LegacyNewDec(math.MaxInt64).MulInt64(100)
 				return params
 			},
-			true,
-			false,
-			0,
-			nil,
+			overrides:       nil,
+			expErr:          true,
+			expVMErr:        false,
+			expectedGasUsed: 0,
+			postCheck:       nil,
 		},
 	}
 
@@ -1047,8 +1125,16 @@ func (s *KeeperTestSuite) TestApplyMessageWithConfig() {
 			)
 			s.Require().NoError(err)
 
-			// Function being tested
-			res, err := s.Network.App.GetEVMKeeper().ApplyMessageWithConfig(s.Network.GetContext(), msg, nil, true, config, txConfig, false)
+			res, err := s.Network.App.GetEVMKeeper().ApplyMessageWithConfig(
+				s.Network.GetContext(),
+				msg,
+				nil,
+				true,
+				config,
+				txConfig,
+				false,
+				tc.overrides,
+			)
 
 			if tc.expErr {
 				s.Require().Error(err)